requirements.txt
github3.py==0.9.5
psycopg2==2.6.1
pytest==2.8.7
PyYAML==3.11
SQLAlchemy==1.0.12
pgpubsub==0.0.5
notify2==0.3
flake8==2.5.4
